Refactor worker to use modular HTML template
